% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.R
\name{sem}
\alias{sem}
\title{Calculate standard error}
\usage{
sem(x)
}
\arguments{
\item{x}{the  vector of numbers}
}
\value{
the standard error
}
\description{
Calculate standard error
}
\examples{
    sem(1:10)
}
